refactor: separate the conversion logic from the generation logic
